% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_prop_z_interval.R
\name{one_prop_z_interval}
\alias{one_prop_z_interval}
\title{Computes a One Proportion Z-Interval}
\usage{
one_prop_z_interval(x, n, confidence_level)
}
\arguments{
\item{x}{The number of observed success in the sample.}

\item{n}{The size of the sample.}

\item{confidence_level}{The confidence level for the interval, often 0.95 is used.}
}
\value{
A list containing the lower and upper bounds of the confidence interval.
}
\description{
This function calculates a confidence interval for a population proportion using a z-distribution.
The confidence interval provides a range of plausible values for the population proportion based on
the proportion observed in a sample. The width of the confidence interval gives an idea of how
uncertain we are about the unknown population proportion. A wider interval implies more uncertainty.
The function assumes that the data follows a binomial distribution and that the sample is random.
}
\examples{
\dontrun{
one_prop_z_interval(5, 10, 0.95)
}
}
