% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_prop_z_test.R
\name{one_prop_z_test}
\alias{one_prop_z_test}
\title{Performs a One Proportion Z-Test}
\usage{
one_prop_z_test(x, n, p0, alternative)
}
\arguments{
\item{x}{The number of observed success in the sample.}

\item{n}{The size of the sample.}

\item{p0}{The hypothesized population proportion.}

\item{alternative}{The the direction of alternative hypothesized. Values can be "less", "greater", or "not_equal".}
}
\value{
A list containing the test statistic and the p-value.
}
\description{
This function implements a One Proportion Z-Test, a statistical hypothesis test used to determine
whether the proportion of successes in a single sample differs significantly from a hypothesized
population proportion. The test assumes that the data follows a binomial distribution and that the
sample is random. The function calculates the test statistic and the p-value. The test statistic
follows a standard normal distribution under the null hypothesis. The p-value represents the
probability that a z-statistic as extreme as the observed one could occur by chance if the null
hypothesis were true.
}
\examples{
\dontrun{
one_prop_z_test(5,10,0.5,"not_equal")
}
}
