% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_interval.R
\name{t_interval}
\alias{t_interval}
\title{Computes a One Sample T-Interval}
\usage{
t_interval(variable1, confidence_level)
}
\arguments{
\item{variable1}{A numeric vector of data values.}

\item{confidence_level}{The confidence level for the interval, often 0.95 is used.}
}
\value{
A list containing the lower and upper bounds of the confidence interval.
}
\description{
This function calculates a confidence interval for a population mean using a t-distribution. The
confidence interval provides a range of plausible values for the population mean based on the mean
observed in a single sample. The width of the confidence interval gives an idea of how uncertain we
are about the unknown population mean. A wider interval implies more uncertainty. The
function assumes that the data follows a normal distribution and that the sample is random.
}
\examples{
\dontrun{
t_interval(c(1,2,3,4,5), 0.95)
}
}
