% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{Performs a One Sample T-Test}
\usage{
t_test(variable1, mu0, alternative)
}
\arguments{
\item{variable1}{A numeric vector of data values.}

\item{mu0}{The hypothesized population mean.}

\item{alternative}{The the direction of alternative hypothesized. Values can be "less", "greater", or "not_equal".}
}
\value{
A list containing the test statistic and the p-value.
}
\description{
This function implements a One Sample T-Test, which is a statistical hypothesis test used to
determine whether the mean of a sample differs significantly from a hypothesized population mean.
The test assumes that the data follows a normal distribution and that the sample is random.
The function calculates the test statistic and the p-value. The test statistic follows a t-distribution
under the null hypothesis. The p-value represents the probability that a t-statistic as extreme as the
observed one could occur by chance if the null hypothesis were true.
}
\examples{
\dontrun{
t_test(c(1,2,3,4,5,6,7,8,9,10), 5, "not_equal")
}
}
