% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_prop_z_interval.R
\name{two_prop_z_interval}
\alias{two_prop_z_interval}
\title{Computes a Two Proportion Z-Interval}
\usage{
two_prop_z_interval(x1, n1, x2, n2, confidence_level)
}
\arguments{
\item{x1}{The number of observed success in sample 1.}

\item{n1}{The size of sample 1.}

\item{x2}{The number of observed success in sample2.}

\item{n2}{The size of sample 2.}

\item{confidence_level}{The confidence level for the interval, often 0.95 is used.}
}
\value{
A list containing the lower and upper bounds of the confidence interval.
}
\description{
This function calculates a confidence interval for the difference between two population proportions
using a z-distribution. The confidence interval provides a range of plausible values for the difference
between two population proportions based on the proportions observed in two independent samples.
The width of the confidence interval gives an idea of how uncertain we are about the unknown difference.
A wider interval implies more uncertainty. The function assumes that the data follows a binomial
distribution and that the samples are independent and random.
}
\examples{
\dontrun{
two_prop_z_interval(5, 10, 50, 100, 0.95)
}
}
