% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_prop_z_test.R
\name{two_prop_z_test}
\alias{two_prop_z_test}
\title{Performs a Two Proportion Z-Test}
\usage{
two_prop_z_test(x1, n1, x2, n2, alternative)
}
\arguments{
\item{x1}{The number of observed success in the sample 1.}

\item{n1}{The size of the sample 1.}

\item{x2}{The number of observed success in the sample 2.}

\item{n2}{The size of the sample 2.}

\item{alternative}{The the direction of alternative hypothesized. Values can be "less", "greater", or "not_equal".}
}
\value{
A list containing the test statistic and the p-value.
}
\description{
This function implements a Two Proportion Z-Test, a statistical hypothesis test used to determine
whether the proportions of successes in two independent samples differ significantly. The test
assumes that the data follows a binomial distribution and that the samples are independent and random.
The function calculates the test statistic and the p-value. The test statistic follows a standard
normal distribution under the null hypothesis. The p-value represents the probability that a
z-statistic as extreme as the observed one could occur by chance if the null hypothesis were true.
}
\examples{
\dontrun{
two_prop_z_test(5, 10, 50, 100, 0.5,"not_equal")
}
}
