% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sample_t_interval.R
\name{two_sample_t_interval}
\alias{two_sample_t_interval}
\title{Computes a Two Sample T-Interval}
\usage{
two_sample_t_interval(variable1, variable2, confidence_level)
}
\arguments{
\item{variable1}{A numeric vector of data values for the first sample.}

\item{variable2}{A numeric vector of data values for the second sample.}

\item{confidence_level}{The confidence level for the interval, often 0.95 is used.}
}
\value{
A list containing the lower and upper bounds of the confidence interval.
}
\description{
This function calculates a confidence interval for the difference between two population means
using a t-distribution. The confidence interval provides a range of plausible values for the
difference between two population means based on the means observed in two independent samples.
The width of the confidence interval gives an idea of how uncertain we are about the unknown difference.
A wider interval implies more uncertainty. The function assumes that the data follows a normal
distribution and that the samples are independent and random.
}
\examples{
\dontrun{
two_sample_t_interval(c(1,2,3,4,5), c(2,4,6,8,10), 0.95)
}
}
