% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sample_t_test.R
\name{two_sample_t_test}
\alias{two_sample_t_test}
\title{Performs a Two Sample T-Test}
\usage{
two_sample_t_test(variable1, variable2, alternative)
}
\arguments{
\item{variable1}{A numeric vector of data values for the first sample.}

\item{variable2}{A numeric vector of data values for the second sample.}

\item{alternative}{The the direction of alternative hypothesized. Values can be "less", "greater", or "not_equal".}
}
\value{
A list containing the test statistic and the p-value.
}
\description{
This function implements a Two Sample T-Test, a statistical hypothesis test used to determine
whether the means of two independent samples differ significantly. It tests the null hypothesis
that the two populations have identical mean values. The test assumes that the data follows a
normal distribution and that the samples are independent and random. The function calculates the
test statistic and the p-value. The test statistic follows a t-distribution under the null hypothesis.
The p-value represents the probability that a t-statistic as extreme as the observed one could occur
by chance if the null hypothesis were true.
}
\examples{
\dontrun{
two_sample_t_test(c(1,2,3,4,5,6,7,8,9,10), c(2,4,6,8,10), "not_equal")
}
}
